FashionSage/docker-requirements.txt
# FastAPI and dependencies
fastapi==0.115.6
uvicorn[standard]==0.35.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.10

# Authentication
bcrypt==4.3.0
PyJWT==2.10.1

# Validation
pydantic[email]==2.9.2
email-validator==2.2.0

# HTTP client
aiohttp==3.12.15

# Additional utilities
python-dotenv==1.0.1